Improve qualifier parsing for R/W status
When the resource is a storage buffer, SPIR-V won't store it's desired qualifier. However, if we have applied a per-usage qualifier it is still valid and can be checked against the current shader we are reflecting on. Further, these qualifiers ARE valid even if SPIR-V doesn't keep them: if used incorrectly, the GLSL won't compile (i.e, we used readonly on a storage buffer we write to)