Remove binarization stuff
It was breaking on weird template errors, and I don't want serialization to really be part of this library anyways. Also, I'm rebuilding this lib and want it to just exist as a silo'd study project for whatever I do next... removing this makes it more reliably what I need now.