Hey so it turns out this is just a compiler now
Committing what I have to track progress, but will be changing things extensively. We need to find the size and alignments of struct members in shaders, and be able to understand how they fit together. We'll use this to optimally arrange things and also tell the frontend all it needs to know about the layout of shader structs ideally